Reactor the code.
