+++ tests/test_nyaml2nxdl.py
@@ -16,68 +16,32 @@
 # limitations under the License.
 #
 """
-Tests for nyaml2nxdl tool
+Tests for the YAML → NXDL conversion direction (nyaml2nxdl).
 """
 
-import filecmp
 import os
-import re
 import sys
 from collections import OrderedDict
-from datetime import datetime
 from pathlib import Path
 
 import lxml.etree as ET
 import pytest
 import yaml
 from click.testing import CliRunner
+from helpers import check_and_replace_latest_copyright, check_file_fresh_baked
 
 from nyaml import cli as nyaml2nxdl
-from nyaml import nyaml2nxdl as nyaml2nxdl_forward_tools
-from nyaml.comment_collector import CommentCollector
-from nyaml.helper import LineLoader, remove_namespace_from_tag
-from nyaml.nyaml2nxdl import get_nxdl_copyright_license, handle_each_part_doc
-
-LATEST_COPYRIGHT_YEAR = f"{datetime.now().year}-{datetime.now().year}"
-LATEST_COPYRIGHT = rf"# Copyright \(C\) {LATEST_COPYRIGHT_YEAR} NeXus International Advisory Committee \(NIAC\)"
-COPYRIGHT_REPLACEMENT = (
-    f"# Copyright (C) 2010-2020 NeXus International Advisory Committee (NIAC)"
-)
-
-
-def check_and_replace_latest_copyright(nxdl_file):
-    """Check if the latest copyright date has been written properly,
-    and replace it accordingly to the test data file.
-    """
-    content = nxdl_file.read_text()
-    generated_copyright = re.findall(LATEST_COPYRIGHT, content, re.DOTALL)
-    assert len(generated_copyright) == 1, (
-        f"No copyright or not correct copyright year found in {nxdl_file}"
-    )
-    content = re.sub(LATEST_COPYRIGHT, COPYRIGHT_REPLACEMENT, content)
-    nxdl_file.write_text(content)
+from nyaml.helper import remove_namespace_from_tag
+from nyaml.nyaml2nxdl import handle_each_part_doc
 
 
 def delete_duplicates(list_of_matching_string):
-    """
-    Delete duplicate from lists
-    """
+    """Delete duplicates from a list while preserving order."""
     return list(dict.fromkeys(list_of_matching_string))
 
 
-def check_file_fresh_baked(test_file):
-    """
-    Get sure that the test file is generated by the converter
-    """
-    path = Path(test_file)
-    timestamp = datetime.fromtimestamp(path.stat().st_mtime).strftime("%d/%m/%Y %H:%M")
-    now = datetime.now().strftime("%d/%m/%Y %H:%M")
-    assert timestamp == now, "xml file not generated"
-
-
 def find_matches(xml_file, desired_matches):
-    """
-        Read xml file and find desired matches. Return a list of two lists in the form:
+    """Read xml file and find desired matches. Return a list of two lists in the form:
     [[matching_line],[matching_line_index]]
     """
     with open(xml_file, encoding="utf-8") as file:
@@ -100,9 +64,8 @@ def find_matches(xml_file, desired_matches):
 
 
 def compare_matches(ref_xml_file, test_yml_file, test_xml_file, desired_matches):
-    """
-        Check if a new xml file is generated
-    and if test xml file is equal to reference xml file
+    """Check if a new xml file is generated
+    and if test xml file is equal to reference xml file.
     """
     # Reference file is read
     ref_matches = find_matches(ref_xml_file, desired_matches)
@@ -114,17 +77,6 @@ def compare_matches(ref_xml_file, test_yml_file, test_xml_file, desired_matches)
     # Test file is read
     test_matches = find_matches(test_xml_file, desired_matches)
     assert test_matches == ref_matches
-
-
-def compare_yaml_content(yaml_dict1, yaml_dict2, test_key_list):
-    for k_val1, k_val2 in zip(yaml_dict1.items(), yaml_dict2.items()):
-        key1, val1 = k_val1
-        key2, val2 = k_val2
-        for ref_key in test_key_list:
-            if key1 == ref_key and key2 == ref_key:
-                assert val1 == val2, f"{ref_key} content is not the same."
-            elif isinstance(val1, dict) and isinstance(val2, dict):
-                compare_yaml_content(val1, val2, test_key_list)
 
 
 def test_links():
@@ -337,61 +289,6 @@ def test_nyaml2nxdl_prohibited_nameType(
     assert expected_error in (str(result.output) + str(result.exception))
 
 
-def test_nxdl2yaml_nameType():
-    """
-    Check the correct handling of the nameType attribute for the direction
-    nxdl->nyaml.
-    """
-    ref_xml_file = "tests/data/nxdl2yaml/allowed_nameType.nxdl.xml"
-    ref_yml_file = "tests/data/nxdl2yaml/ref_allowed_nameType.yaml"
-    test_yml_file = "tests/data/nxdl2yaml/allowed_nameType_parsed.yaml"
-
-    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_xml_file])
-    assert result.exit_code == 0
-    check_file_fresh_baked(test_yml_file)
-
-    result = filecmp.cmp(ref_yml_file, test_yml_file, shallow=False)
-
-    assert result, "Ref YML and parsed YML don't have the same structure!"
-    os.remove(test_yml_file)
-    sys.stdout.write("Test on xml -> yml nameType okay.\n")
-
-    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_xml_file])
-    assert result.exit_code == 0
-    check_file_fresh_baked(test_yml_file)
-
-    result = filecmp.cmp(ref_yml_file, test_yml_file, shallow=False)
-    assert result, (
-        "Ref YML and parsed YML\
-has not the same structure!!"
-    )
-    os.remove(test_yml_file)
-    sys.stdout.write("Test on xml -> yml doc formatting okay.\n")
-
-
-def test_nxdl2yaml_doc_format_and_nxdl_part_as_comment():
-    """
-    This test for two reasons:
-        1. In test-1 an nxdl file with all kind of doc formats are translated
-    to yaml to check if they are correct.
-        2. In test-2: Check the nxdl that comes at the end of yaml file as comment.
-    """
-    ref_xml_file = "tests/data/Ref_NXentry.nxdl.xml"
-    ref_yml_file = "tests/data/Ref_NXentry.yaml"
-    test_yml_file = "tests/data/Ref_NXentry_parsed.yaml"
-    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_xml_file])
-    assert result.exit_code == 0
-    check_file_fresh_baked(test_yml_file)
-
-    result = filecmp.cmp(ref_yml_file, test_yml_file, shallow=False)
-    assert result, (
-        "Ref YML and parsed YML\
-has not the same structure!!"
-    )
-    os.remove(test_yml_file)
-    sys.stdout.write("Test on xml -> yml doc formatting okay.\n")
-
-
 def test_file_line_error():
     """
     In this test the yaml file line in the error message is tested.
@@ -488,137 +385,6 @@ def test_enumerations_nyaml2nxdl():
     sys.stdout.write("Test on open/closed enumerations okay.\n")
 
 
-def test_nxdl2yaml_enumerations():
-    """
-    Check the correct handling of enumerations (closed and open ones) for the direction
-    nxdl->nyaml.
-    """
-    ref_xml_file = "tests/data/nxdl2yaml/enumerations.nxdl.xml"
-    ref_yml_file = "tests/data/nxdl2yaml/ref_enumerations.yaml"
-    test_yml_file = "tests/data/nxdl2yaml/enumerations_parsed.yaml"
-
-    result = CliRunner().invoke(
-        nyaml2nxdl.launch_tool, ["--do-not-store-nxdl", str(ref_xml_file)]
-    )
-    assert result.exit_code == 0
-    check_file_fresh_baked(test_yml_file)
-
-    result = filecmp.cmp(ref_yml_file, test_yml_file, shallow=False)
-
-    assert result, "Ref YML and parsed YML don't have the same structure!"
-    os.remove(test_yml_file)
-    sys.stdout.write("Test on xml -> yml doc formatting okay.\n")
-
-
-def test_xml_parsing():
-    """
-        In this test an xml file in converted to yml and then back to xml.
-    The xml trees of the two files are then compared.
-    """
-    ref_xml_file = "tests/data/Ref_NXellipsometry.nxdl.xml"
-    test_yml_file = "tests/data/Ref_NXellipsometry_parsed.yaml"
-    test_xml_file = "tests/data/Ref_NXellipsometry_parsed.nxdl.xml"
-    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_xml_file])
-    assert result.exit_code == 0
-    check_file_fresh_baked(test_yml_file)
-    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [test_yml_file])
-    assert result.exit_code == 0
-    check_file_fresh_baked(test_xml_file)
-
-    test_tree = ET.parse(test_xml_file)
-    test_tree_flattened = {i.tag.split("}", 1)[1] for i in test_tree.iter()}
-
-    ref_tree = ET.parse(ref_xml_file)
-    ref_tree_flattened = {i.tag.split("}", 1)[1] for i in ref_tree.iter()}
-
-    assert test_tree_flattened == ref_tree_flattened, (
-        "Ref XML and parsed XML\
-has not the same tree structure!!"
-    )
-    os.remove(test_xml_file)
-    os.remove(test_yml_file)
-    sys.stdout.write("Test on xml -> yml -> xml okay.\n")
-
-
-def test_yml_parsing():
-    """In this test an xml file in converted to yml and then back to xml.
-    The xml trees of the two files are then compared.
-    """
-    ref_yml_file = "tests/data/Ref_NXellipsometry.yaml"
-    test_xml_file = "tests/data/Ref_NXellipsometry.nxdl.xml"
-    test_yml_file = "tests/data/Ref_NXellipsometry_parsed.yaml"
-    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_yml_file])
-    assert result.exit_code == 0
-    check_file_fresh_baked(test_xml_file)
-    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [test_xml_file])
-    assert result.exit_code == 0
-    check_file_fresh_baked(test_yml_file)
-
-    test_yml_tree = nyaml2nxdl_forward_tools.yml_reader(test_yml_file)
-
-    ref_yml_tree = nyaml2nxdl_forward_tools.yml_reader(ref_yml_file)
-
-    assert list(test_yml_tree) == list(ref_yml_tree), (
-        "Ref YML and parsed YML \
-has not the same root entries!!"
-    )
-    os.remove("tests/data/Ref_NXellipsometry_parsed.yaml")
-    sys.stdout.write("Test on yml -> xml -> yml okay.\n")
-
-
-def test_yml_consistency_comment_parsing():
-    """Test comments parsing from yaml. Convert 'yaml' input file to '.nxdl.xml' and
-    '.nxdl.xml' to '.yaml'
-    """
-
-    ref_yml_file = "tests/data/Ref_NXcomment.yaml"
-    test_yml_file = "tests/data/Ref_NXcomment_consistency.yaml"
-
-    result = CliRunner().invoke(
-        nyaml2nxdl.launch_tool, ["--check-consistency", ref_yml_file]
-    )
-    assert result.exit_code == 0, (
-        f"Exception: {result.exception}, \nExecution Info:{{result.exc_info}}"
-    )
-    with open(ref_yml_file, encoding="utf-8") as ref_yml:
-        loader = LineLoader(ref_yml)
-        ref_loaded_yaml = loader.get_single_data()
-    ref_comment_blocks = CommentCollector(ref_yml_file, ref_loaded_yaml)
-    ref_comment_blocks.extract_all_comment_blocks()
-
-    with open(test_yml_file, encoding="utf-8") as test_yml:
-        loader = LineLoader(test_yml)
-        test_loaded_yaml = loader.get_single_data()
-    test_comment_blocks = CommentCollector(test_yml_file, test_loaded_yaml)
-    test_comment_blocks.extract_all_comment_blocks()
-
-    for reference_comment, test_comment in zip(ref_comment_blocks, test_comment_blocks):
-        assert reference_comment == test_comment, "Comment is not consistent."
-
-    os.remove(test_yml_file)
-
-
-def test_conversion():
-    """
-    Test conversion of NXentry
-    """
-    root = Path(__file__).parent / "data" / "NXentry.nxdl.xml"
-    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [str(root)])
-    assert result.exit_code == 0
-    # Replace suffixes
-    yaml = root.parent / Path(root.with_suffix("").stem + "_parsed.yaml")
-    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [str(yaml)])
-    assert result.exit_code == 0
-    new_root = yaml.with_suffix(".nxdl.xml")
-    with open(root, encoding="utf-8") as tmp_f:
-        root_content = tmp_f.readlines()
-    with open(new_root, encoding="utf-8") as tmp_f:
-        new_root_content = tmp_f.readlines()
-    assert root_content == new_root_content
-    Path.unlink(yaml)
-    Path.unlink(new_root)
-
-
 def test_yaml2nxdl_doc():
     """To test the doc style from yaml to nxdl."""
     pwd = Path(__file__).parent
@@ -655,34 +421,6 @@ def test_yaml2nxdl_doc():
     compare_nxdl_doc(ref_nxdl, out_nxdl)
 
     Path.unlink(out_doc_file)
-
-
-def test_nxdl2yaml_doc():
-    """To test the doc style from nxdl to yaml."""
-
-    pwd = Path(__file__).parent
-    nxdl_file = pwd / "data/doc_nxdl2yaml.nxdl.xml"
-    ref_yaml = pwd / "data/ref_doc_nxdl2yaml.yaml"
-    parsed_yaml_file = pwd / "data/doc_nxdl2yaml_parsed.yaml"
-
-    result = CliRunner().invoke(
-        nyaml2nxdl.launch_tool, ["--do-not-store-nxdl", str(nxdl_file)]
-    )
-
-    if result.exit_code != 0:
-        Path.unlink(parsed_yaml_file)
-
-    assert result.exit_code == 0, "Error in converter execution."
-
-    with (
-        open(ref_yaml, encoding="utf-8") as yaml1,
-        open(parsed_yaml_file, encoding="utf-8") as yaml2,
-    ):
-        yaml_dict1 = LineLoader(yaml1).get_single_data()
-        yaml_dict2 = LineLoader(yaml2).get_single_data()
-
-    compare_yaml_content(yaml_dict1, yaml_dict2, ["doc"])
-    Path.unlink(parsed_yaml_file)
 
 
 def test_yaml2nxdl_no_tabs(tmp_path):
@@ -743,78 +481,6 @@ def test_copyright_license_new_yaml(tmp_path):
     # Check if the latest copyright year is written
     check_and_replace_latest_copyright(output)
     """
-
-
-def test_check_copyright_license_in_full_modification_yaml_cycle(tmp_path):
-    pwd = Path(__file__).parent
-    nxdl_file = pwd / "data" / "Ref_NXentry_License.nxdl.xml"
-    yaml_file = tmp_path / "Ref_NXentry_License_parsed.yaml"
-    modified_yaml_gen = tmp_path / "Ref_NXentry_License_modified.yaml"
-    modified_yaml_ref = pwd / "data" / "Ref_NXentry_License_modified.yaml"
-    latest_nxdl = tmp_path / "Ref_NXentry_License_modified.nxdl.xml"
-
-    result = CliRunner().invoke(
-        nyaml2nxdl.launch_tool, [str(nxdl_file), "--output-file", str(yaml_file)]
-    )
-    assert result.exit_code == 0, (
-        f"Error in converter execution input file {nxdl_file}."
-    )
-    content = yaml_file.read_text()
-    find_pattern = r"my nice doc string in root level, line 2."
-    replace_pattern = "my nice doc string in root level, line 2. Modified."
-    updated_content = re.sub(find_pattern, replace_pattern, content)
-    modified_yaml_gen.write_text(updated_content)
-    # Compare two yaml and modified yaml
-    with (
-        open(modified_yaml_gen, encoding="utf-8") as gen_yaml,
-        open(modified_yaml_ref, encoding="utf-8") as ref_yaml,
-    ):
-        gen_yaml_dict = LineLoader(gen_yaml).get_single_data()
-        ref_yaml_dict = LineLoader(ref_yaml).get_single_data()
-    compare_yaml_content(gen_yaml_dict, ref_yaml_dict, ["doc"])
-
-    # Convert modified yaml to nxdl
-    result = CliRunner().invoke(
-        nyaml2nxdl.launch_tool,
-        [str(modified_yaml_gen), "--output-file", str(latest_nxdl)],
-    )
-
-    gen_license_text = get_nxdl_copyright_license(latest_nxdl)
-    original_license_text = get_nxdl_copyright_license(nxdl_file)
-    assert gen_license_text == original_license_text, "License text is not correct."
-
-
-def test_check_copyright_license_in_modified_yaml(tmp_path):
-    """While converting the modified yaml to nxdl the license text should
-    come from stored nxdl file.
-    """
-    pwd = Path(__file__).parent
-    yaml_file = pwd / "data/Ref_NXentry.yaml"
-    modified_yaml = tmp_path / "NXentry_modified.yaml"
-    output = tmp_path / "NXentry_modified.nxdl.xml"
-
-    content = yaml_file.read_text()
-    find_pattern = r"my nice doc string in root level, line 2."
-    replace_pattern = "my nice doc string in root level, line 2. Modified."
-    updated_content = re.sub(find_pattern, replace_pattern, content)
-    modified_yaml.write_text(updated_content)
-
-    result = CliRunner().invoke(
-        nyaml2nxdl.launch_tool, [str(modified_yaml), "--output-file", str(output)]
-    )
-    assert result.exit_code == 0, (
-        f"Error in converter execution input file {modified_yaml}."
-    )
-
-    expected_text = (
-        r"Copyright \(C\) 2010-2020 NeXus International Advisory Committee \(NIAC\)"
-    )
-
-    license_text = get_nxdl_copyright_license(output)
-    assert license_text, "License text not found in nxdl file."
-    text_list = re.findall(expected_text, license_text, re.DOTALL)
-
-    assert len(text_list) == 1, "License text is not correct."
 
 
 @pytest.mark.parametrize(
@@ -923,6 +589,8 @@ def test_forward_conversion(test_input):
     result = runner.invoke(nyaml2nxdl.launch_tool, [test_yml_input_file])
     assert result.exit_code == 0
 
+    # check_and_replace_latest_copyright(Path(test_xml_output_file))
+
     with open(test_xml_output_file, encoding="utf-8") as logfile:
         log = logfile.readlines()
     with open(ref_xml_output_file, encoding="utf-8") as reference_file:
@@ -930,31 +598,3 @@ def test_forward_conversion(test_input):
     assert log == ref
 
     os.remove(test_xml_output_file)
-
-
-@pytest.mark.parametrize(
-    "test_input",
-    [
-        ("NXdimensionsType"),
-    ],
-)
-def test_backward_conversion(test_input):
-    """
-    Tests if the conversion of specific test files from NXDL to YAML results as expected.
-    Expected output files shall have the corresponding name with the prefix Ref_.
-    """
-    prefix = "tests/data/nxdl2yaml/"
-    test_xml_input_file = prefix + test_input + ".nxdl.xml"
-    test_yml_output_file = prefix + test_input + "_parsed.yaml"
-    ref_yml_output_file = prefix + "Ref_" + test_input + ".yaml"
-    runner = CliRunner()
-    result = runner.invoke(nyaml2nxdl.launch_tool, [test_xml_input_file])
-    assert result.exit_code == 0
-
-    with open(test_yml_output_file, encoding="utf-8") as logfile:
-        log = logfile.readlines()
-    with open(ref_yml_output_file, encoding="utf-8") as reference_file:
-        ref = reference_file.readlines()
-    assert log == ref
-
-    os.remove(test_yml_output_file)
